refactor: move ecstore endpoint layout owners
