+++ crates/ecstore/src/layout/mod.rs
@@ -5,5 +5,7 @@
 //! file moves happen.
 
 pub(crate) mod disks_layout;
+pub(crate) mod endpoint;
+pub(crate) mod endpoints;
 pub(crate) mod format;
 pub(crate) mod set_layout;
